# backend/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "novela",
    "summary": "Estado, capítulos y manifiestos; lanzamiento de novelas",
    "version": "0.1.0"
  },
  "paths": {
...
    "/novelas/{slug}/estado": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Estado",
        "operationId": "estado_novelas__slug__estado_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estado"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/config": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Config",
        "operationId": "config_novelas__slug__config_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/escaleta": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Escaleta",
        "operationId": "escaleta_novelas__slug__escaleta_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Escaleta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/checkpoint": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Checkpoint",
        "operationId": "checkpoint_novelas__slug__checkpoint_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Checkpoint"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Checkpoint Novelas  Slug  Checkpoint Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/runs": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Runs",
        "operationId": "runs_novelas__slug__runs_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Manifest"
                  },
                  "title": "Response Runs Novelas  Slug  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/runs/{run_id}/log": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Log",
        "description": "El tramo de `harness.log` desde el byte `desde`, que el panel encadena con `hasta` (D48).\nUn run sin log todavía responde un tramo vacío: solo existe entre el manifiesto y la primera\nlínea.",
        "operationId": "log_novelas__slug__runs__run_id__log_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^r-[0-9]{8}-[0-9]{4}$",
              "title": "Run Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desde"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TramoDeLog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/runs/{run_id}": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Manifiesto",
        "operationId": "manifiesto_novelas__slug__runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^r-[0-9]{8}-[0-9]{4}$",
              "title": "Run Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/portada": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Portada",
        "operationId": "portada_novelas__slug__portada_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La portada, sin texto",
            "content": {
              "image/jpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/novelas/{slug}/metricas": {
      "get": {
        "tags": [
          "novelas"
        ],
        "summary": "Metricas",
        "operationId": "metricas_novelas__slug__metricas_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InformeDeCostes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Checkpoint": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "capitulo": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Capitulo"
          },
          "cursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "run_id": {
            "type": "string",
            "pattern": "^r-[0-9]{8}-[0-9]{4}$",
            "title": "Run Id"
          },
          "version_canon": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Version Canon"
          },
          "version_plan": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Version Plan"
          },
          "capitulos_sha256": {
            "additionalProperties": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$"
            },
            "type": "object",
            "title": "Capitulos Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capitulo",
          "cursor",
          "run_id",
          "version_canon",
          "version_plan",
          "capitulos_sha256"
        ],
        "title": "Checkpoint",
        "description": "`checkpoints/NN.json` y `latest.json`: cursor, versiones y run_id, no la base (RF-20).\n`capitulos_sha256` es el sello de los capítulos cerrados hasta N (RF-35)."
      },
      "Config": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "idea_semilla": {
            "type": "string",
            "minLength": 1,
            "title": "Idea Semilla"
          },
          "parametros_obra": {
            "$ref": "#/components/schemas/ParametrosObra"
          },
          "parametros_sistema": {
            "$ref": "#/components/schemas/ParametrosSistema"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idea_semilla",
          "parametros_obra",
          "parametros_sistema"
        ],
        "title": "Config"
      },
...
      "Cursor": {
        "properties": {
          "capitulo": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Capitulo"
          },
          "fase": {
            "type": "string",
            "enum": [
              "escritura",
              "revision",
              "registro",
              "cerrado"
            ],
            "title": "Fase"
          },
          "ultimo_paso": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "briefing",
                  "escritor",
                  "validar",
                  "continuista",
                  "editor-estilo",
                  "lector-suspense",
                  "cronista",
                  "aplicar-delta",
                  "checkpoint"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Paso"
          },
          "intento": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Intento"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capitulo",
          "fase",
          "ultimo_paso",
          "intento"
        ],
        "title": "Cursor"
      },
...
      "EntradaConocimiento": {
        "properties": {
          "hecho": {
            "type": "string",
            "pattern": "^hec-[0-9]{3}$",
            "title": "Hecho"
          },
          "desde_capitulo": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Desde Capitulo"
          },
          "cita": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cita"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hecho",
          "desde_capitulo"
        ],
        "title": "EntradaConocimiento",
        "description": "Misma entrada para `conocimiento` (por personaje) y `conocimiento_lector` (plana)."
      },
...
      "EntradaTemporal": {
        "properties": {
          "escena": {
            "type": "string",
            "pattern": "^esc-[0-9]{2,3}-[0-9]+$",
            "title": "Escena"
          },
          "capitulo": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Capitulo"
          },
          "inicio": {
            "type": "string",
            "title": "Inicio"
          },
          "duracion_min": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duracion Min"
          },
          "cita": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cita"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "escena",
          "capitulo",
          "inicio",
          "duracion_min"
        ],
        "title": "EntradaTemporal"
      },
      "Escaleta": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "actos": {
            "items": {
              "$ref": "#/components/schemas/Acto"
            },
            "type": "array",
            "minItems": 1,
            "title": "Actos"
          },
          "puntos_de_giro": {
            "$ref": "#/components/schemas/PuntosDeGiro"
          },
          "curva_tension_objetivo": {
            "items": {
              "type": "integer",
              "maximum": 10.0,
              "minimum": 1.0
            },
            "type": "array",
            "title": "Curva Tension Objetivo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actos",
          "puntos_de_giro",
          "curva_tension_objetivo"
        ],
        "title": "Escaleta",
        "description": "Se valida con `context={\"num_capitulos\": N}`: la curva tiene un valor por capítulo."
      },
      "Estado": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "cursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "linea_temporal": {
            "items": {
              "$ref": "#/components/schemas/EntradaTemporal"
            },
            "type": "array",
            "title": "Linea Temporal",
            "default": []
          },
          "personajes": {
            "patternProperties": {
              "^per-[a-z0-9-]+$": {
                "$ref": "#/components/schemas/EstadoPersonaje"
              }
            },
            "type": "object",
            "title": "Personajes",
            "default": {}
          },
          "conocimiento": {
            "patternProperties": {
              "^per-[a-z0-9-]+$": {
                "items": {
                  "$ref": "#/components/schemas/EntradaConocimiento"
                },
                "type": "array"
              }
            },
            "type": "object",
            "title": "Conocimiento",
            "default": {}
          },
          "relaciones": {
            "items": {
              "$ref": "#/components/schemas/Relacion"
            },
            "type": "array",
            "title": "Relaciones",
            "default": []
          },
          "objetos": {
            "items": {
              "$ref": "#/components/schemas/Objeto"
            },
            "type": "array",
            "title": "Objetos",
            "default": []
          },
          "libro_de_hechos": {
            "items": {
              "$ref": "#/components/schemas/Hecho"
            },
            "type": "array",
            "title": "Libro De Hechos",
            "default": []
          },
          "hilos": {
            "items": {
              "$ref": "#/components/schemas/Hilo"
            },
            "type": "array",
            "title": "Hilos",
            "default": []
          },
          "pistas": {
            "patternProperties": {
              "^pis-[0-9]{3}$": {
                "$ref": "#/components/schemas/EstadoPista"
              }
            },
            "type": "object",
            "title": "Pistas",
            "default": {}
          },
          "conocimiento_lector": {
            "items": {
              "$ref": "#/components/schemas/EntradaConocimiento"
            },
            "type": "array",
            "title": "Conocimiento Lector",
            "default": []
          },
          "tension_real": {
            "items": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10.0,
                  "minimum": 1.0
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Tension Real",
            "default": []
          },
          "metricas": {
            "$ref": "#/components/schemas/Metricas",
            "default": {
              "palabras_totales": 0,
              "desviacion_vs_plan": 0.0
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cursor"
        ],
        "title": "Estado"
      },
      "EstadoPersonaje": {
        "properties": {
          "ubicacion": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^esc-[a-z][a-z0-9-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ubicacion"
          },
          "estado_fisico": {
            "type": "string",
            "title": "Estado Fisico"
          },
          "estado_emocional": {
            "type": "string",
            "title": "Estado Emocional"
          },
          "condicion": {
            "type": "string",
            "enum": [
              "viva",
              "muerta",
              "desaparecida"
            ],
            "title": "Condicion"
          },
          "objetivo_activo": {
            "type": "string",
            "title": "Objetivo Activo"
          },
          "ultima_aparicion": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Ultima Aparicion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ubicacion",
          "estado_fisico",
          "estado_emocional",
          "condicion",
          "objetivo_activo",
          "ultima_aparicion"
        ],
        "title": "EstadoPersonaje"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hecho": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^hec-[0-9]{3}$",
            "title": "Id"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          },
          "capitulo": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Capitulo"
          },
          "cita": {
            "type": "string",
            "title": "Cita"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "texto",
          "capitulo",
          "cita"
        ],
        "title": "Hecho"
      },
...
      "InformeDeCostes": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "sesion": {
            "type": "string",
            "title": "Sesion"
          },
          "generado": {
            "type": "string",
            "format": "date-time",
            "title": "Generado"
          },
          "pasos": {
            "items": {
              "$ref": "#/components/schemas/ConsumoDePaso"
            },
            "type": "array",
            "title": "Pasos"
          },
          "total": {
            "$ref": "#/components/schemas/TotalDeCostes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "slug",
          "sesion",
          "generado",
          "pasos",
          "total"
        ],
        "title": "InformeDeCostes"
      },
...
      "Manifest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0.0",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "run_id": {
            "type": "string",
            "pattern": "^r-[0-9]{8}-[0-9]{4}$",
            "title": "Run Id"
          },
          "capitulo": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Capitulo"
          },
          "fase": {
            "type": "string",
            "enum": [
              "arranque",
              "capitulo"
            ],
            "title": "Fase",
            "default": "capitulo"
          },
          "creado": {
            "type": "string",
            "title": "Creado"
          },
          "sha_commit": {
            "type": "string",
            "title": "Sha Commit"
          },
          "version_recetas": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Version Recetas"
          },
          "version_canon": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Version Canon"
          },
          "version_plan": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Version Plan"
          },
          "sucio": {
            "type": "boolean",
            "title": "Sucio",
            "default": false
          },
          "hashes_claude": {
            "additionalProperties": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$"
            },
            "type": "object",
            "title": "Hashes Claude",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "capitulo",
          "creado",
          "sha_commit",
          "version_recetas",
          "version_canon",
          "version_plan"
        ],
        "title": "Manifest",
        "description": "`runs/<run_id>/manifest.json`: lo que permite atribuir un cambio de calidad a un cambio\nconcreto (validators.md §4.7). Las versiones son hashes de contenido, no números. `fase`\nsepara el run de arranque —arquitecto y trazador— de los runs de capítulo (spec 0003)."
      },
...
      "Objeto": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^obj-[0-9]{3}$",
            "title": "Id"
          },
          "poseedor": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^per-[a-z0-9-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poseedor"
          },
          "ubicacion": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^esc-[a-z][a-z0-9-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ubicacion"
          },
          "capitulo_intro": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Capitulo Intro"
          },
          "relevancia": {
            "type": "string",
            "enum": [
              "alta",
              "media",
              "baja"
            ],
            "title": "Relevancia"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "poseedor",
          "ubicacion",
          "capitulo_intro",
          "relevancia"
        ],
        "title": "Objeto"
      },
...
      "Relacion": {
        "properties": {
          "de": {
            "type": "string",
            "pattern": "^per-[a-z0-9-]+$",
            "title": "De"
          },
          "a": {
            "type": "string",
            "pattern": "^per-[a-z0-9-]+$",
            "title": "A"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "intensidad": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Intensidad"
          },
          "desde": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Desde"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "de",
          "a",
          "tipo",
          "intensidad",
          "desde"
        ],
        "title": "Relacion",
        "description": "Arista de la evolución; `intensidad` es cuantitativa y no es la `tension` del canon."
      },
...
      "TramoDeLog": {
        "properties": {
          "desde": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Desde"
          },
          "hasta": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Hasta"
          },
          "tamano": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tamano"
          },
          "modificado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modificado"
          },
          "lineas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lineas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "desde",
          "hasta",
          "tamano",
          "modificado",
          "lineas"
        ],
        "title": "TramoDeLog",
        "description": "Líneas completas de `runs/<run_id>/harness.log` desde el byte `desde`. `hasta` es el\nsiguiente `desde`: el cliente encadena sin que el servidor guarde nada. Sin log, `tamano` 0,\n`modificado` None y `hasta` = `desde`. No es documento del workspace: sin schema_version."
      },
...
    }
  }
}